+++ CHANGELOG.md
@@ -6,4 +6,20 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.0.5] - 2022-10-16
+
+- Simple Bug Fixes
+
+## [0.0.4] - 2022-10-16
+
+### Added 
+- A new sidebar with different functionalites
+- A window is opened only when run command is clicked.
+- Compile, test, publish, clean and funding of account with faucet can be done in sidebar itself.
+- Clean and fund with faucet are new functionalities
+- When the argument types are entered, a dropdown menu is show as you type to prevent mistakes.
+
+### Removed 
+- Extra functionalites like compile, test and publish are removed from the window. These can be done from sidebar itself.
+
 - Initial release